Function Pointer Arguments Must NOT be instrumented for Offset2Pointer Conversion and TaintedPointer Check

+++ include/checkcbox_extensions.h
@@ -17,7 +17,7 @@
 #pragma TLIB_SCOPE on
 int c_isPointerToTaintedMem(void* pointer);
 
-void* c_fetch_pointer_from_offset(const unsigned long pointer_offset);
+void* c_fetch_pointer_from_offset(const unsigned int pointer_offset);
 
 int c_isTaintedPointerToTaintedMem(void* pointer);
